Support for external generators in RTE Model (continue)
make some common code accessible from different tools
remove duplicates
